Fix PasswordKey/BrainKey key generation.
We can't use base Graphene objects as is, because they
create keys with wrong prefixes ("GPH"). Maybe this should
be fixed the other way around.

This also adds `get_blind_private` to BrainKey class. As the
name suggests, that is the correct method to generate private keys
for blind accounts.